add endpoint for volume stats
Presently this only includes the volume size.

[#109469982]

+++ volume/driver.go
@@ -5,7 +5,7 @@ package volume
 type Driver interface {
 	CreateVolume(path string) error
 	DestroyVolume(path string) error
-	GetVolumeSize(path string) (uint64, error)
+	GetVolumeSize(path string) (uint, error)
 
 	CreateCopyOnWriteLayer(path string, parent string) error
 }
